Added missing capabilities to WMAppManifest
Even though the example application doesn't need any special permissions
to use the phone's different capabilities, Visual Studio considers the
project to be made with an old version of Windows Phone dev tools if the
Capabilities tag is empty.

To resolve this problem I added Medialib capability inside the tag.