Accurate sliding and bug fixes
Earlier the sliding was done by touching anywhere inside the control. It
was actually problematic because in the middle of the control there's
optionally a textbox which shows the current value of the slider. If the
developer allows, tapping the textbox opens a digit prompt which allows
input of the slider value. Whenever this was tapped, the slider most
likely moved too and it was very confusing.

Now the control checks if the touch event is occuring outisde the middle
ellipse and inside the control bounds (which is elliptical). It means
that now the sliding and changing the value via textbox is much better
and it even works with weirdly shaped sliders (really high but short in
width for example).

I also added a new page to the example app showcasing strangely shaped
RadialSliders and how they work also.